Coin system
Added a sprite for the coins. Added the script for the coinpickup where it dissapears when the player collides with it. Added the script for player that when it collides with a coin it will increment the money vealue of the player with the value of the coined that got picked up. Added a bunch of coins on the scene to test it up.