Enemy sprite + fireball + animation
Changed the enemy sprite. Added an idle animation for it. Changed the enemy bullet sprite to a fireball. Changed the way the enemy shoots (doesn't shoot anymore from a gun but instead it creates the fireball) so I removed the gun completly from the enemy and just kept the firepoint. The enemyWeapon script got moved to the enemy prefab. Added the posibility for the fireball to instantiate 180 degrees on x axis when the enemy is facing left (added in the enemysenses)